Inventory/warehouse-sync/requirements.txt
aiohttp==3.9.5
prometheus-client==0.20.0
kafka-python==2.0.2
sqlalchemy==2.0.30
asyncpg==0.29.0
pydantic==2.7.3
PyYAML==6.0.1
apscheduler==3.10.1

# Dev dependencies
pytest==8.2.2
pytest-asyncio==0.23.6
aioresponses==0.7.6
